+++ .gitignore
@@ -172,3 +172,8 @@ cython_debug/
 
 # PyPI configuration file
 .pypirc
+
+
+
+# DO NOT PUSH TOKEN
+.token_cache
